% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvar_fevd.R
\name{diff_fevd}
\alias{diff_fevd}
\title{In construction...}
\usage{
diff_fevd(
  fevd_each,
  response = NULL,
  baseline = c(1, 1),
  plot = TRUE,
  bar_args = list(),
  facet_args = list()
)
}
\arguments{
\item{facet_args}{}
}
\description{
In construction...
}
