% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvar_irf.R
\name{diff_irf}
\alias{diff_irf}
\title{In construction...}
\usage{
diff_irf(
  irf_each,
  response,
  impulse,
  baseline = 1,
  line_args = list(),
  ribbon_args = list(fill = NA, color = "blue", linetype = 2),
  hline_args = list(),
  facet_args = list()
)
}
\arguments{
\item{facet_args}{}
}
\description{
In construction...
}
