% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdiff_values.R
\name{ggdiff_values}
\alias{ggdiff_values}
\title{Title}
\usage{
ggdiff_values(
  data,
  splits_each,
  index,
  cols = NULL,
  mod_names = NULL,
  ...,
  line_args = list(),
  facet_args = list()
)
}
\arguments{
\item{facet_args}{}
}
\description{
Title
}
