% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvar_var.R
\name{ggvar_select_patch}
\alias{ggvar_select_patch}
\title{In construction...}
\usage{
ggvar_select_patch(select_each, model_labs = names(select_each), ...)
}
\arguments{
\item{...}{}
}
\description{
In construction...
}
