% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_values_shaded.R
\name{ggvar_values_shaded}
\alias{ggvar_values_shaded}
\title{Title}
\usage{
ggvar_values_shaded(
  data,
  splits_each,
  index,
  cols = NULL,
  splits_names = NULL,
  ...,
  colors = NULL,
  line_args = list(),
  facet_args = list()
)
}
\arguments{
\item{colors}{}
}
\description{
Title
}
