% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvar_acf.R
\name{mapvar_acf}
\alias{mapvar_acf}
\title{Calculate ACF for each data split or model}
\usage{
mapvar_acf(x_each, cols_each = NULL, plot = TRUE, ...)
}
\arguments{
\item{x_each}{A list of datasets or VAR models to get variables or residuals
from.}

\item{cols_each}{A list of columns to consider for each data split. Defaults to everything.}

\item{plot}{If a ggplot should be returned (the default), instead of the raw
data. Soon to be removed in favor of a \link[base]{plot} method.}

\item{...}{Further arguments passed to \link[stats]{acf}.}
}
\value{
A list of datasets with the ACF information for each split/model, or
a ggplot, if \code{plot = TRUE}
}
\description{
Calculate the Auto-Correlation function for each data split of each VAR  model
residuals.
}
\examples{
data <- EuStockMarkets
data_each <- split_data(data, list(c(1, 620), c(621, 1240), c(1241)))
mapvar_acf(data_each, plot = FALSE)

}
