% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_acf.R, R/mapvar_acf.R
\name{mapvar_ccf}
\alias{mapvar_ccf}
\title{Title}
\usage{
mapvar_ccf(data_each, cols = NULL, plot = TRUE, patch = FALSE, ...)

mapvar_ccf(data_each, cols = NULL, plot = TRUE, patch = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title

Title
}
