% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvar_exogen.R
\name{mapvar_exogen}
\alias{mapvar_exogen}
\title{Create exogenous variables for VAR models}
\usage{
mapvar_exogen(data_each, fun = NULL, args_each = NULL, k = length(data_each))
}
\arguments{
\item{data_each}{A list of data splits to build exogenous variables with
each.}

\item{fun}{A function that takes a data split, and a set of additional
arguments, and returns a dataset with exogenous variables.}

\item{args_each}{A list with sets of exogenous arguments to be passed to
\code{fun} for each split. It can be a single item or a list with unit
length to pass the same arguments for each split.}

\item{k}{The number of splits. Should be \code{length(data_each)}.}
}
\description{
This function creates a set of exogenous variables for each data split. It
takes a custom function and a set of arguments for each split. This can be
used to supply the \code{exogen} argument of functions in the "vars"
package.
}
