% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvar_fun.R, R/vardiff_map.R
\name{mapvar_fun}
\alias{mapvar_fun}
\title{Title}
\usage{
mapvar_fun(
  x_each,
  what,
  ...,
  pretty = "none",
  pretty_cond = NULL,
  cols = NULL,
  patch_args = list(),
  broom_args = list(),
  star_args = list()
)

mapvar_fun(
  x_each,
  what,
  ...,
  pretty = "none",
  pretty_cond = NULL,
  cols = NULL,
  patch_args = list(),
  broom_args = list(),
  star_args = list()
)
}
\arguments{
\item{star_args}{}
}
\description{
Title

Title
}
