% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvar_var.R
\name{mapvar_select}
\alias{mapvar_select}
\title{In construction...}
\usage{
mapvar_select(
  data_each,
  p_each,
  cols_each = NULL,
  ...,
  criteria = c("AIC", "HQ", "SC", "FPE"),
  exogen_fun = NULL,
  exogen_args = NULL,
  star = FALSE,
  star_args = list()
)
}
\arguments{
\item{...}{}
}
\description{
In construction...
}
