% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_var.R
\name{mapvar_varselect}
\alias{mapvar_varselect}
\title{Title}
\usage{
mapvar_varselect(
  data_each,
  p_each,
  cols_each = NULL,
  ...,
  criteria = c("AIC", "HQ", "SC", "FPE"),
  exogen_fun = NULL,
  exogen_args = NULL,
  star = FALSE,
  star_args = list()
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
