% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vardiff-package.R
\docType{package}
\name{vardiff-package}
\alias{vardiff}
\alias{vardiff-package}
\title{vardiff: Creating and Comparing Multiple VAR Models}
\description{
A package to create and work with multiple VAR models. Split data, vizualise the splitted variables, create and diagnose var models for each split, then compare results based on IRF's, FEVD's, CF's, and causality tests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ricardo-semiao/vardiff}
  \item \url{https://ricardo-semiao.github.io/vardiff/}
  \item Report bugs at \url{https://github.com/ricardo-semiao/vardiff/issues}
}

}
\author{
\strong{Maintainer}: Ricardo Semião \email{ricardo.semiao@outlook.com}

}
\keyword{internal}
