% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvar_causality.R
\name{mapvar_causality}
\alias{mapvar_causality}
\title{In construction...}
\usage{
mapvar_causality(
  mod_each,
  cause = NULL,
  ...,
  vcov_fun = NULL,
  vcov_args = list(),
  star = FALSE,
  star_args = list(),
  star_names = list(),
  star_round = 3
)
}
\arguments{
\item{mod_each}{}
}
\description{
In construction...
}
