% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvar_acf.R
\name{mapvar_ccf}
\alias{mapvar_ccf}
\title{Calculate CCF for each data split or model}
\usage{
mapvar_ccf(x_each, cols = NULL, plot = TRUE, patch = FALSE, ...)
}
\arguments{
\item{x_each}{A list of datasets or VAR models to get variables or residuals
from.}

\item{cols}{Columns to consider.}

\item{plot}{If a ggplot should be returned (the default), instead of the raw
data. Soon to be removed in favor of a \link[base]{plot} method.}

\item{patch}{If \code{plot = TRUE}, should the plots be wrapped
(\link[patchwork]{wrap_plots}) together or not (the default)?}

\item{...}{Further arguments passed to \link[stats]{acf}.}
}
\value{
A list of datasets with the ACF information for each split/model (if
\code{plot = FALSE}), a list of ggplots (else if \code{patch = FALSE}), or a
patched ggplot.
}
\description{
Calculate the Cross-Correlation function for each data split of each VAR
model residuals.
}
\examples{
data <- EuStockMarkets
data_each <- split_data(data, list(c(1, 620), c(621, 1240), c(1241)))
mapvar_ccf(data_each, plot = FALSE)

}
