% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_exogen.R, R/mapvar_exogen.R
\name{mapvar_exogen}
\alias{mapvar_exogen}
\title{Title}
\usage{
mapvar_exogen(data_each, fun = NULL, args = NULL, k = length(data_each))

mapvar_exogen(data_each, fun = NULL, args = NULL, k = length(data_each))
}
\arguments{
\item{args}{}

\item{k}{}
}
\description{
Title

Title
}
