% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvar_var.R
\name{mapvar_var}
\alias{mapvar_var}
\title{Title}
\usage{
mapvar_var(
  data_each,
  p_each,
  cols_each = NULL,
  exogen_fun = NULL,
  exogen_args = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
