% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vardiff-package.R
\docType{package}
\name{vardiff-package}
\alias{vardiff}
\alias{vardiff-package}
\title{vardiff: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ricardo-semiao/vardiff}
  \item \url{https://ricardo-semiao.github.io/vardiff/}
  \item Report bugs at \url{https://github.com/ricardo-semiao/vardiff/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
